Split r2c
